more aesthetic changes

+++ src/components/NewsletterEditor/index.js
@@ -1,4 +1,5 @@
 import React, { Component, useState } from 'react';
+import { Button } from 'reactstrap';
 import { withFirebase } from '../Firebase';
 
 
@@ -23,8 +24,13 @@ class NewsletterEditor extends Component {
 
     render() {
         return (
-            <div>
-                <input className="btn btn-outline-primary" onClick={this.toggleNewsletterEditor} value={this.state.newsletterButtonText} type="button"/>
+            <div style={{margin:"auto", width:"100%"}}>
+                <input
+                    onClick={this.toggleNewsletterEditor}
+                    value={this.state.newsletterButtonText}
+                    type="button"
+                    className="btn btn-outline-primary"
+                />
                 { !this.state.newsletterIsHidden ? <NewsletterMenu firebase={this.props.firebase} /> : null }
             </div>
         );
@@ -43,7 +49,7 @@ const NewsletterMenu = ({firebase}) => {
         month = `0${month}`
     }
     let today=`${year}-${month}-${day}`
-    
+
     const [titleOfNewsletterToBeDeleted, setTitleOfNewsletterToBeDeleted] = useState('');
     const [dateOfNewsletterToBeDeleted, setDateOfNewsletterToBeDeleted] = useState('');
 
@@ -58,43 +64,69 @@ const NewsletterMenu = ({firebase}) => {
 
     const handleNewsletterDeleteForm = (event) => {
         event.preventDefault();
-        console.log(`Title of newsletter to be deleted: ${titleOfNewsletterToBeDeleted}`);
-        console.log(`Date of newsletter to be deleted: ${dateOfNewsletterToBeDeleted}`);
         //ADD DATABASE STUFF HERE
         //ASK COREY ABOUT HOW TO GO ABOUT THIS
-        firebase.deleteNewsLetters(`${titleOfNewsletterToBeDeleted}`,`${dateOfNewsletterToBeDeleted}`)
+        if(window.confirm("Are you sure that you want to delete this Newsletter?")) {
+            firebase.deleteNewsLetters(`${titleOfNewsletterToBeDeleted}`, `${dateOfNewsletterToBeDeleted}`)
+        }
     }
 
     const handleNewsletterAddForm = (event) => {
         event.preventDefault();
-        console.log(`Title of newsletter to be added: ${titleOfNewsletterToBeAdded}`);
-        console.log(`Content of newsletter to be added: ${contentOfNewsletterToBeAdded}`);
-        console.log(`Linked Video of newsletter to be added: ${linkedVideoOfNewsletterToBeAdded}`);
-        console.log(`Date of newsletter to be added: ${dateOfNewsletterToBeAdded}`);
-        //ADD DATABASE STUFF HERE
+        if(titleOfNewsletterToBeAdded == '') {
+            window.alert("A title must be provided");
+        }
+            //ADD DATABASE STUFF HERE
         //ASK COREY ABOUT HOW TO GO ABOUT THIS
-        firebase.insertNewsLetters(`${contentOfNewsletterToBeAdded}`,`${dateOfNewsletterToBeAdded}`,`${linkedVideoOfNewsletterToBeAdded}`,`${titleOfNewsletterToBeAdded}`)
+        else {
+            firebase.insertNewsLetters(`${contentOfNewsletterToBeAdded}`, `${dateOfNewsletterToBeAdded}`, `${linkedVideoOfNewsletterToBeAdded}`, `${titleOfNewsletterToBeAdded}`)
+            window.alert("Newsletter created!");
+        }
     }
 
     return(
-        <div>
+        <div >
             <br/>
-            <h5>Add a newsletter</h5>
-            <form onSubmit={handleNewsletterAddForm} name="addNewsletter">
-                Title: <input type="text" name="title" placeholder="Title" size="50" value={titleOfNewsletterToBeAdded} onChange={updateState(setTitleOfNewsletterToBeAdded)}/><br/>
-                Content: <br/><textarea rows="5" cols="50" name="content" wrap="soft" placeholder="Insert newsletter text content here..." value={contentOfNewsletterToBeAdded} onChange={updateState(setContentOfNewsletterToBeAdded)}/><br/>
-                Linked Video (Enter "NONE" if no video to link): <br/><input type="text" name="linkedVideo" size="50" placeholder="URL of linked YouTube video" value={linkedVideoOfNewsletterToBeAdded} onChange={updateState(setLinkedVideoOfNewsletterToBeAdded)}/><br/>
-                Date: <input type="date" name="date" value={dateOfNewsletterToBeAdded} onChange={updateState(setDateOfNewsletterToBeAdded)}/>
-                <br/><input type="submit" value="Submit"/>
-            </form>
+            <div className="card card-1 shadow" >
+                <h4 style={{margin:"auto"}}>Add a newsletter</h4>
+                <br/>
+                <form onSubmit={handleNewsletterAddForm} name="addNewsletter"  style={{margin:"auto"}}>
+                    Title: <input type="text" name="title" placeholder="Title" size="50" value={titleOfNewsletterToBeAdded} onChange={updateState(setTitleOfNewsletterToBeAdded)}/>
+                    <br/>
+                    Content: <br/>
+                    <textarea rows="5" cols="50" name="content" wrap="soft" placeholder="Insert newsletter text content here..." value={contentOfNewsletterToBeAdded} onChange={updateState(setContentOfNewsletterToBeAdded)}/>
+                    <br/>
+                    Linked Video (Enter "NONE" if no video to link): <br/>
+                    <br/>
+                    <input type="text" name="linkedVideo" size="50" placeholder="URL of linked YouTube video" value={linkedVideoOfNewsletterToBeAdded} onChange={updateState(setLinkedVideoOfNewsletterToBeAdded)}/>
+                    <br/>
+                    <br/>
+                    Date: <input type="date" name="date" value={dateOfNewsletterToBeAdded} onChange={updateState(setDateOfNewsletterToBeAdded)}/>
+                    <br/>
+                    <br/>
+                    <Button color="primary" onClick={(e) => handleNewsletterAddForm(e)}>Insert</Button>
+                </form>
+                <br/>
+            </div>
             <br/>
-            <h5>Delete a newsletter</h5>
-            Enter the title and date of the newsletter to be deleted:
-            <form onSubmit={handleNewsletterDeleteForm} id="deleteNewsletter">
-                Title: <input type="text" name="title" placeholder="Title" size="50" value={titleOfNewsletterToBeDeleted} onChange={updateState(setTitleOfNewsletterToBeDeleted)}/><br/>
-                Date:<input type="date" name="date" value={dateOfNewsletterToBeDeleted} onChange={updateState(setDateOfNewsletterToBeDeleted)}/><br/>
-                <input type="submit" value="Submit"/>
-            </form>
+            <div className="card card-1 shadow">
+                <h4 style={{margin:"auto"}} >Delete a newsletter</h4>
+                <br/>
+                <span style={{margin:"auto"}} >Enter the title and date of the newsletter to be deleted:</span>
+                <br/>
+                <form onSubmit={handleNewsletterDeleteForm} id="deleteNewsletter"  style={{margin:"auto"}}>
+                    Title: <input type="text" name="title" placeholder="Title" size="50" value={titleOfNewsletterToBeDeleted} onChange={updateState(setTitleOfNewsletterToBeDeleted)}/>
+                    <br/>
+                    <br/>
+                    Date:<input type="date" name="date" value={dateOfNewsletterToBeDeleted} onChange={updateState(setDateOfNewsletterToBeDeleted)}/>
+                    <br/>
+                    <br/>
+                    <Button color="danger" onClick={(e) => handleNewsletterDeleteForm(e)}>Delete</Button>
+                    <br/>
+                    <br/>
+                </form>
+            </div>
+            <br/>
         </div>
     );
 }
